trigger document changes
For this document, in a few places you need to change "that IP address" to "that the IP address" and also "requests were originated" to "requests originated" because these changes make the sentence flow much better. Also, you need to change "requests originated" to "requests that originated" and in a couple places change "requests were originated" to "requests originated" because it makes more sense like this. The last change that needs to be put in is "to patch" to "patching". Otherwise, the document looks very good!