api discovery document changes
This document had a lot of spelling and grammar errors like changing "and variety" to "and a variety, changing "the" to "they", changing "vectors" to "vector", changing "on matching" to "of matching", changing "target" to "targets", and etc. As you can see most of the changes I showed and more and just a lot of simple grammar and spelling errors that need to be fixed that the sentences sound and flow better.