api discover document changes
This document had little changes but the changes were the same because you need to add "the" to some of the sentences. the changes were "in API" to "in the API", then "as endpoint" to "as the endpoint", and "re-calculate risk" to "re-calculate the risk". Overall the document was very good!