configure kubernetes document changes
This document had some changes that needed to be made like changing "of Wallarm" to "of the Wallarm" and "of NGINX" to "of the NGINX" because these two need "the" in the middle to make the sentence flow better. Then I made "value" plural to "values", then "to Wallarm" to "the Wallarm" then "points the" to "points to the", the "requests" to "requested". These changes are basic grammar and spelling changes that make the sentences flow. Overall the document looks good!